third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-callstack.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    function innerTestFunction()
    {
        timeout1();
    }
    setTimeout(innerTestFunction, 0);
}

function timeout1()
{
    debugger;
    requestAnimationFrame(animFrame1);
    var id = setInterval(innerInterval1, 0);
    function innerInterval1()
    {
        clearInterval(id);
        interval1();
    }
}

function animFrame1()
{
    debugger;
    setTimeout(timeout2, 0);
    requestAnimationFrame(animFrame2);
}

function interval1()
{
    debugger;
}

function timeout2()
{
    debugger;
}

function animFrame2()
{
    debugger;
    function longTail0()
    {
        timeout3();
    }
    var funcs = [];
    for (var i = 0; i < 20; ++i)
        funcs.push("function longTail" + (i + 1) + "() { setTimeout(longTail" + i + ", 0); };");
    funcs.push("setTimeout(longTail" + i + ", 0);");
    eval(funcs.join("\n"));
}

function timeout3()
{
    debugger;
}

var test = function()
{
    var totalDebuggerStatements = 6;
    InspectorTest.runAsyncCallStacksTest(totalDebuggerStatements);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests asynchronous call stacks in debugger.
</p>

</body>
</html>
